repair sample generation for combinatoric with minimum trials
When a minimum-trials constraint extends the number of trials to be
not a multiple of the crossing size, the uniform combinatoric sampler
was reoprting the right total number of trials, but it was not
correctly varying the non-crossing-sized tail, and so it was reporting
duplicates.

The repair here is in the way the sequence number -> permutation
calculation is performed so that it's compatible with prefixes.

There are additional changes intended as setup for a change in the way
crossed, derived factors with complex windows are treated, but that
change is not complete or enabled here. It would have been cleaner to
keep those changes separate, but I discovered the minimum-trials
problems working toward the other change, and this commit is a
checkpoint in case that other change doesn't work out.