change combinatoric sampling of designs with complex crossed factors
This change reduces the search space for rejection sampling with
complex derived factors by many orders of magnitude. However, it's
complicated, and the reductions are not enough orders.

For example, PadmalaPessoa2011 does down from a seach space of 2^463
to 2^253, while Stroop_simple goes from 2^122 to 2^106. That's not
enough to make rejection sampling useful.